Divide ObjectM;anagerContainer Knife getting a bit stiff...
