+++ GamesToGo.Desktop/Project/Elements/IHasElements.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GamesToGo.Desktop.Project.Elements
+{
+    public interface IHasElements<T> where T : ProjectElement
+    {
+        void AddElement(int ID);
+    }
+}
